Move AbacusCoreChecker to deploy
