+++ typescript/infra/package.json
@@ -1,11 +1,11 @@
 {
   "name": "@abacus-network/infra",
   "description": "Infrastructure utilities for the Abacus Network",
-  "version": "0.2.1-alpha7",
+  "version": "0.2.1-alpha11",
   "dependencies": {
-    "@abacus-network/core": "^0.2.1-alpha7",
-    "@abacus-network/deploy": "^0.2.1-alpha7",
-    "@abacus-network/sdk": "^0.2.1-alpha7",
+    "@abacus-network/core": "^0.2.1-alpha11",
+    "@abacus-network/deploy": "^0.2.1-alpha11",
+    "@abacus-network/sdk": "^0.2.1-alpha11",
     "@aws-sdk/client-iam": "^3.74.0",
     "@aws-sdk/client-kms": "3.48.0",
     "@aws-sdk/client-s3": "^3.74.0",
