--- typescript/infra/src/core/index.ts
@@ -1,1 +1,0 @@
-export { AbacusCoreChecker, CoreViolationType } from './check';
