+++ typescript/infra/src/utils/utils.ts
@@ -160,13 +160,3 @@ export function writeJSON(directory: string, filename: string, obj: any) {
     JSON.stringify(obj, null, 2),
   );
 }
-
-// Returns a \ b
-// Taken from https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Set#implementing_basic_set_operations
-export function setDifference<T>(a: Set<T>, b: Set<T>) {
-  const diff = new Set(a);
-  for (const element of b) {
-    diff.delete(element);
-  }
-  return diff;
-}
