+++ typescript/sdk/src/core/events.ts
@@ -1,5 +1,5 @@
-import type { ProcessEvent } from '@abacus-network/core/types/contracts/Inbox';
-import type { DispatchEvent } from '@abacus-network/core/types/contracts/Outbox';
+import type { ProcessEvent } from '@abacus-network/core/dist/contracts/Inbox';
+import type { DispatchEvent } from '@abacus-network/core/dist/contracts/Outbox';
 
 import { Annotated } from '../events';
